requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.11.0
PyMySQL==1.1.0
python-dateutil==2.8.2
reportlab==4.0.9
drf-spectacular==0.26.5
gunicorn==21.2.0

# Logging & Monitoring
sentry-sdk==1.39.1
python-json-logger==2.0.7
django-extensions==3.2.3

# Database & Performance
django-db-conn-pool==1.3.3

# Rate Limiting
django-ratelimit==4.1.0

# Testing & Code Quality (optional, for development)
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
